Set clipboard in nvim and rebind tmux copy paste
- Update neovim.nix
- Updated tmux.nix

For neovim, set the clipboard to unnamedplus makes it so operations like
yank will push to my system clipboard when used with a terminal
like foot. Does not work in gnome-terminal.

For tmux, set copy-mode key bindings to start selection with v and
copy selection with y, just like in vim.